refactor : custom hook 분리 및 appkey 관리
